+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+## v5.33.0
+
+- Add `alert_events_url` to alert source resource and data source
+
 ## v5.32.0
 
 - Add `incident_maintenance_window` resources to create maintenance windows